backend/requirements.txt
python-telegram-bot==13.3
emoji==1.2.0
requests==2.24.0
beautifulsoup4==4.9.3
lxml==4.6.2
